Move interop root importing calldata encoding to api helpers
