+++ api/Cargo.toml
@@ -24,6 +24,8 @@ evm_interpreter = { path = "../evm_interpreter", default-features = false }
 ruint = { version = "1.12.3", default-features = false, features = ["alloc"] }
 alloy = { version = "=1", default-features = false, features = ["eip712", "consensus", "rpc-types", "signer-local", "dyn-abi", "network"] }
 zksync_os_interface = { workspace = true }
+alloy-sol-types = "1"
+
 
 [features]
 # By default, use the risc_v_simulator with delegation (which is faster).
